Handle custom sreg attribute namespaces
A namespace attribute was currently only supported for ax attributes. If
a namespace was used with sreg attributes the openid provider was unable
to find the attributes.

This was the case when using the openid provider with an Atlassian Crowd
Server (respectively Atlassian CrowdID), which was using a sreg
namespace.

The default namespace remains 'sreg'. If the response from the
OpenID server contains an item with the key 'openid_ns_<namespace>' and
the value 'http://openid.net/extensions/sreg/1.1' the namespace is
extracted and used to retrieve sreg attributes.